Add basic support for chained certificates
Sync in the charm-helpers fix which configures the Apache2
openstack_https_frontend to allow SSL Chained Certificates.

Publicly issued SSL certificates are in recent years generally
issued from an intermediate root certificate, creating a trust
chain from the server certificate through the intermediate root
up to the ultimately trusted root CA certificate in the system
SSL store. Most clients (including firefox, wget and many python
SSL libraries) require the server to send the full SSL
certificate chain to the client to correctly validate this
arrangement (notably, Google Chrome does not. Beware of that
if testing this yourself).

apache2 >= 2.4.8 (xenial+) already handles this allowing you
to simply concatenate all 3 certificates in order (from leaf
to root) into the SSLCertificateFile and it automatically
handles sending this full chain to the client.

apache2 <= 2.4.7 (trusty-) requires this to be configured
separately with the SSLCertificateChainFile directive, refer
to the upstream documentation:

http://httpd.apache.org/docs/2.4/mod/mod_ssl.html#sslcertificatechainfile

We configure the SSLCertificateChainFile to use the same certificate
file as SSLCertificateFile. This works well allowing the same
configuration of simply concatenating all 3 certificates be used
on any version.

The presence of the server certificate in the chain file does not appear
to cause any problems. Because we re-use the existing certificate file,
no code changes are required in charms to support this.

Change-Id: If84c4bc310ae98c84017ef8e2de8c595c3f34d3a
Partial-Bug: #1484489